requirements.txt
# app/requirements.txt
fastapi==0.70.1
uvicorn[standard]==0.22.0
sqlalchemy==1.4.48
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
email-validator==1.1.3
aiosqlite==0.19.0
jinja2==3.1.4
requests==2.31.0
xmltodict==0.13.0
python-multipart==0.0.7
PyYAML==6.0.1
pydantic==1.10.13
itsdangerous==2.1.2
pycryptodome==3.19.0
PyJWT==2.8.0
boto3==1.34.0